+++ app/src/main/java/com/iti/mobile/triporganizer/data/repository/notes/NoteRepositoryRoomImp.java
@@ -32,7 +32,7 @@ public class NoteRepositoryRoomImp implements NotesRepository {
     }
 
     @Override
-    public LiveData<List<Note>> getNotesForTrip(int tripId, String userId) {
+    public LiveData<List<Note>> getNotesForTrip(long tripId, String userId) {
         return notesRoom.getAllNotes(tripId);
     }
 }
