+++ app/src/main/java/com/iti/mobile/triporganizer/data/room/dao/NoteDao.java
@@ -25,5 +25,5 @@ public interface NoteDao {
     void deleteNote(Note note);
 
     @Query("SELECT * FROM notes WHERE tripId = :tripId")
-    LiveData<List<Note>> getAllNote(int tripId);
+    LiveData<List<Note>> getAllNote(long tripId);
 }
